+++ changelog.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keepa Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/purecloudlabs/softphone-vendor-headsets/compare/v3.0.1...HEAD)
+# [Unreleased](https://github.com/purecloudlabs/softphone-vendor-headsets/compare/v3.0.2...HEAD)
+
+# [v3.0.2](https://github.com/purecloudlabs/softphone-vendor-headsets/compare/v3.0.1...v3.0.2)
+## Fixed
+* [STREAM-1800](https://inindca.atlassian.net/browse/STREAM-1800) - Fixed Yealink WebHID report ID resolution for devices (like the UH38) that nest telephony input reports in a child HID collection.
+
+## Added
+* [STREAM-1707](https://inindca.atlassian.net/browse/STREAM-1707) - Added UI integration tests for the react app.
 
 # [v3.0.1](https://github.com/purecloudlabs/softphone-vendor-headsets/compare/v3.0.0...v3.0.1)
 ## Fixed